Add VsockConnectionManagerCommon and make it generic over VirtIOSocketManager
The new VsockConnectionManagerCommon type allows using a common implementation
for a subset of the methods VsockConnectionManager provides. Adding a new type
for the common methods is preferred over the trait approach taken for
VirtIOSocket mthods because the trait runs into split borrow issues when
borrowing the fields of VsockConnectionManager. This commit avoids breaking the
existing API by redefining VsockConnectionManager in terms of
VsockConnectionManagerCommon and making methods delegate to the implementations
in VsockConnectionManagerCommon.