Add client_id to DeviceTransport and DeviceHal::dma_map
This is what I pushed to gerrit [1] to support per-VM vsock devices, but the
problem is ultimately that the dma hal uses PhysAddr to represent virtio-msg
"bus addresses." This could be a general problem for other virtio transports so
the proper fix may be to allow the dma hal to define PhysAddr, with a sensible
default. This way dma HALs for virtio-msg over FFA could use a bus addresses
plus the FFA VM ID when specifying what to map in.

[1] https://android-review.googlesource.com/c/platform/external/rust/android-crates-io/+/3487653/2..3