net: split header struct to legacy/nonlegacy
Size of header struct depends on whether we're using legacy VIRTIO or in
legacy case, if MRG_RXBUF feature was negotiated. This is a problem
because in many places we rely on the header size which was hardcoded.

Encode whether we're using the legacy layout and decide on runtime which
header struct layout to use. If using legacy header, the missing fields
are filled with zeros.

This is hidden from the API user.

Signed-off-by: Manos Pitsidianakis <manos.pitsidianakis@linaro.org>