Implement ack_interrupt_raw for PciTransport
This method is an alternative to `Transport::ack_interrupt` which
does not require a mutable reference to the entire
`PciTransport`. This is useful for acknowleding interrupts in
cases where it may be difficult to safely obtain a mutable
reference to the transport (e.g. in an interrupt handler while
another thread is using the same PciTransport embedded in a
VirtIOSocket.