Add VsockConnectionManager::get_transport
This commit adds a `get_transport` method for the driver-side
`VsockConnectionManager` type to allow using `PciTransport::ack_interrupt_raw`. It
requires boxing or putting the VsockConnectionManager in an Arc to ensure the
transport pointer it returns has a stable address. Since only the driver side is
currently supported, code relying on the `VsockManager` trait to be generic over
both the device and driver sides must get the transport pointer in the
non-generic part (e.g. when creating the Arc<VsockConnectionManager>). This
commit also adds tests showing the most straightforward ways to use the Pin API.