Add ack_interrupt method to VsockManager trait
This commit adds an ack_interrupt method which takes a pointer to
the VsockManager trait. It may be called on the driver side in
the same way that VirtIOSocket's equivalent method can, but the
device side will panic if this method is called. The method is
implemented in the vsock manager trait rather than only for the
driver-specific VsockConnectionManager to simplify code that
tries to be generic over which side of the vsock connection is
calling it.